Refactoring into CLIP plugin
