+++ allenact_plugins/clip_plugin/extra_environment.yml
@@ -1,0 +1,13 @@
+channels:
+  - defaults
+  - conda-forge
+  - pytorch
+dependencies:
+  - pytorch=1.7.1
+  - torchvision
+  - cudatoolkit=11.0
+  - pip:
+      - ftfy
+      - regex
+      - tqdm
+      - git+https://github.com/openai/CLIP.git
