+++ allenact_plugins/clip_plugin/extra_requirements.txt
@@ -1,0 +1,6 @@
+pytorch==1.7.1
+torchvision
+ftfy
+regex
+tqdm
+clip @ git+git://github.com/openai/clip@e184f608c5d5e58165682f7c332c3a8b4c1545f2#egg=clip
