+++ allenact_plugins/clip_plugin/objectnav_models.py
