+++ projects/objectnav_baselines/experiments/clip/objectnav_zeroshot_mixin_clip_gru.py
@@ -5,14 +5,16 @@ import numpy as np
 import torch.nn as nn
 
 from allenact.base_abstractions.preprocessor import Preprocessor
-from allenact.embodiedai.preprocessors.resnet import ClipResNetPreprocessor
-from allenact.embodiedai.preprocessors.text_encoders import ClipTextPreprocessor
+from allenact_plugins.clip_plugin.clip_preprocessors import (
+    ClipResNetPreprocessor,
+    ClipTextPreprocessor
+)
 from allenact.embodiedai.sensors.vision_sensors import RGBSensor
 from allenact.utils.experiment_utils import Builder
 from allenact_plugins.ithor_plugin.ithor_sensors import GoalObjectTypeThorSensor
 from allenact_plugins.robothor_plugin.robothor_tasks import ObjectNavTask
 from projects.objectnav_baselines.experiments.objectnav_base import ObjectNavBaseConfig
-from projects.objectnav_baselines.models.clip_models import CLIPObjectNavActorCritic
+from allenact_plugins.clip_plugin.objectnav_models import CLIPObjectNavActorCritic
 
 
 class ObjectNavZeroShotMixInClipGRUConfig(ObjectNavBaseConfig):
