+++ projects/objectnav_baselines/experiments/robothor/clip/objectnav_robothor_rgb_clipresnet50gru_ddppo.py
@@ -1,9 +1,9 @@
-from allenact.embodiedai.preprocessors.resnet import ClipResNetPreprocessor
+from allenact_plugins.clip_plugin.clip_preprocessors import ClipResNetPreprocessor
 from allenact_plugins.ithor_plugin.ithor_sensors import (
     RGBSensorThor,
     GoalObjectTypeThorSensor,
 )
-from projects.objectnav_baselines.experiments.objectnav_mixin_clipresnetgru import (
+from projects.objectnav_baselines.experiments.clip.objectnav_mixin_clipresnetgru import (
     ObjectNavMixInClipResNetGRUConfig,
 )
 from projects.objectnav_baselines.experiments.objectnav_mixin_ddppo import (
